task/gen-334 changes implemented
